add: fragment description start

+++ app/src/main/java/com/example/kinopoisk/descriptionFragment/models/modelForDescription/Rating.kt
@@ -1,9 +1,9 @@
 package com.example.kinopoisk.descriptionFragment.models.modelForDescription
 
 data class Rating(
-    val await: Int,
+    val await: Double,
     val filmCritics: Double,
     val imdb: Double,
     val kp: Double,
-    val russianFilmCritics: Int
+    val russianFilmCritics: Double
 )
